refactor importExport functionality into modules
