+++ src/modules/importExport/importExportEntities/entities/contentItems.ts
@@ -1,10 +1,10 @@
 import { ContentItemContracts, ManagementClient } from "@kontent-ai/management-sdk";
 import chalk from "chalk";
 
-import { logInfo, LogOptions } from "../../../log.js";
-import { zip } from "../../../utils/array.js";
-import { serially } from "../../../utils/requests.js";
-import { ReplaceReferences } from "../../../utils/types.js";
+import { logInfo, LogOptions } from "../../../../log.js";
+import { zip } from "../../../../utils/array.js";
+import { serially } from "../../../../utils/requests.js";
+import { ReplaceReferences } from "../../../../utils/types.js";
 import { getRequired } from "../../import/utils.js";
 import { EntityDefinition, ImportContext } from "../entityDefinition.js";
