+++ tests/unit/import/replaceRichTextReferences.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from "@jest/globals";
 
-import { replaceRichTextReferences } from "../../../src/commands/importExportEntities/entities/utils/richText";
+import { replaceRichTextReferences } from "../../../src/modules/importExport/importExportEntities/entities/utils/richText";
 
 describe("replaceRichTextReferences", () => {
   it("replaces asset references as internal id", () => {
